+++ src/main/kotlin/org/zowe/explorer/config/ws/ui/WSUsernameColumn.kt
@@ -16,49 +16,31 @@ package org.zowe.explorer.config.ws.ui
 
 import com.intellij.util.ui.ColumnInfo
 import org.zowe.explorer.common.message
-import org.zowe.explorer.common.ui.ErrorableTableCellRenderer
+import org.zowe.explorer.config.connect.ui.renderer.UsernameColumnRenderer
 import org.zowe.explorer.config.ws.WorkingSetConfig
-import java.awt.Component
-import javax.swing.JTable
-import javax.swing.table.DefaultTableCellRenderer
 import javax.swing.table.TableCellRenderer
 
-private val NO_USERNAME_MESSAGE = message("configurable.ws.tables.ws.username.error.empty")
 
 /**
  * Class which represents working set username column in working set table model
  */
 class WSUsernameColumn<WSConfig : WorkingSetConfig>(
-  private val getUsername: (WSConfig) -> String?,
-  private val renderUsername: (WSConfig) -> String? = { getUsername(it) }
+  private val getUsername: (WSConfig) -> String,
+  private val checkIsZoweConfig: (WSConfig) -> Boolean
 ) : ColumnInfo<WSConfig, String>(message("configurable.ws.tables.ws.username.name")) {
 
   /**
    * Overloaded getter method. Gets the username from crudable by connection config uuid
    */
   override fun valueOf(item: WSConfig): String {
-    return getUsername(item) ?: NO_USERNAME_MESSAGE
+    return getUsername(item)
   }
 
   /**
    * Overloaded getter method. Gets the renderer
    */
   override fun getRenderer(item: WSConfig): TableCellRenderer {
-    return ErrorableTableCellRenderer(
-      errorMessage = message("configurable.ws.tables.ws.url.error.empty.tooltip"),
-      renderer = object: DefaultTableCellRenderer() {
-        override fun getTableCellRendererComponent(
-          table: JTable?, value: Any?, isSelected: Boolean, hasFocus: Boolean, row: Int, column: Int
-        ): Component {
-          val formedUsername = renderUsername(item)
-          return super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column).apply {
-            text = formedUsername
-          }
-        }
-      }
-    ) {
-      getUsername(item) == null
-    }
+    return UsernameColumnRenderer(checkIsZoweConfig(item))
   }
 
   /**
@@ -66,13 +48,6 @@ class WSUsernameColumn<WSConfig : WorkingSetConfig>(
    */
   override fun isCellEditable(item: WSConfig?): Boolean {
     return false
-  }
-
-  /**
-   * Gets the UI tooltip of the username column when mouse is hovered
-   */
-  override fun getTooltipText(): String {
-    return message("configurable.ws.tables.ws.username.tooltip")
   }
 
 }