+++ src/main/kotlin/org/zowe/explorer/explorer/actions/RenameAction.kt
@@ -27,15 +27,7 @@ import org.zowe.explorer.dataops.attributes.RemoteMemberAttributes
 import org.zowe.explorer.dataops.attributes.RemoteUssAttributes
 import org.zowe.explorer.dataops.content.synchronizer.checkFileForSync
 import org.zowe.explorer.dataops.operations.RenameOperation
-import org.zowe.explorer.explorer.ui.ExplorerTreeNode
-import org.zowe.explorer.explorer.ui.FileExplorerView
-import org.zowe.explorer.explorer.ui.FileLikeDatasetNode
-import org.zowe.explorer.explorer.ui.LibraryNode
-import org.zowe.explorer.explorer.ui.RenameDialog
-import org.zowe.explorer.explorer.ui.UssDirNode
-import org.zowe.explorer.explorer.ui.UssFileNode
-import org.zowe.explorer.explorer.ui.cleanCacheIfPossible
-import org.zowe.explorer.explorer.ui.getExplorerView
+import org.zowe.explorer.explorer.ui.*
 import org.zowe.explorer.telemetry.NotificationsService
 import org.zowe.explorer.vfs.MFVirtualFile
 
@@ -89,7 +81,7 @@ class RenameAction : AnAction() {
           node.parent?.cleanCacheIfPossible(cleanBatchedQuery = true)
         }
         .onFailure {
-          NotificationsService.getService().notifyError(it, project)
+          NotificationsService.errorNotification(it, project)
         }
     }
   }