+++ src/test/kotlin/org/zowe/explorer/explorer/actions/AllocateDatasetActionTestSpec.kt
@@ -161,7 +161,9 @@ class AllocateDatasetActionTestSpec : WithApplicationShouldSpec({
         every { viewMock.mySelectedNodesData } returns selectedNodesData
         every { workingSetMock.name } returns "test"
         every { workingSetMock.uuid } returns "test"
-        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns mockk<ConnectionConfig>()
+        val connectionConfig = mockk<ConnectionConfig>()
+        every { connectionConfig.uuid } returns "fake_uuid"
+        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns connectionConfig
         val dataOpsManager = DataOpsManager.getService() as TestDataOpsManagerImpl
         dataOpsManager.testInstance = object : TestDataOpsManagerImpl() {
           override fun <R : Any> performOperation(operation: Operation<R>, progressIndicator: ProgressIndicator): R {
@@ -238,7 +240,9 @@ class AllocateDatasetActionTestSpec : WithApplicationShouldSpec({
         every { viewMock.mySelectedNodesData } returns selectedNodesData
         every { workingSetMock.name } returns "test"
         every { workingSetMock.uuid } returns "test"
-        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns mockk<ConnectionConfig>()
+        val connectionConfig = mockk<ConnectionConfig>()
+        every { connectionConfig.uuid } returns "fake_uuid"
+        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns connectionConfig
         val dataOpsManager = DataOpsManager.getService() as TestDataOpsManagerImpl
         dataOpsManager.testInstance = object : TestDataOpsManagerImpl() {
           override fun <R : Any> performOperation(operation: Operation<R>, progressIndicator: ProgressIndicator): R {
@@ -312,7 +316,9 @@ class AllocateDatasetActionTestSpec : WithApplicationShouldSpec({
         every { viewMock.mySelectedNodesData } returns selectedNodesData
         every { workingSetMock.name } returns "test"
         every { workingSetMock.uuid } returns "test"
-        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns mockk<ConnectionConfig>()
+        val connectionConfig = mockk<ConnectionConfig>()
+        every { connectionConfig.uuid } returns "fake_uuid"
+        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns connectionConfig
         val dataOpsManager = DataOpsManager.getService() as TestDataOpsManagerImpl
         dataOpsManager.testInstance = object : TestDataOpsManagerImpl() {
           override fun <R : Any> performOperation(operation: Operation<R>, progressIndicator: ProgressIndicator): R {
@@ -383,7 +389,9 @@ class AllocateDatasetActionTestSpec : WithApplicationShouldSpec({
         every { viewMock.mySelectedNodesData } returns selectedNodesData
         every { workingSetMock.name } returns "test"
         every { workingSetMock.uuid } returns "test"
-        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns mockk<ConnectionConfig>()
+        val connectionConfig = mockk<ConnectionConfig>()
+        every { connectionConfig.uuid } returns "fake_uuid"
+        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns connectionConfig
         val dataOpsManager = DataOpsManager.getService() as TestDataOpsManagerImpl
         dataOpsManager.testInstance = object : TestDataOpsManagerImpl() {
           override fun <R : Any> performOperation(operation: Operation<R>, progressIndicator: ProgressIndicator): R {
@@ -456,7 +464,9 @@ class AllocateDatasetActionTestSpec : WithApplicationShouldSpec({
         every { viewMock.mySelectedNodesData } returns selectedNodesData
         every { workingSetMock.name } returns "test"
         every { workingSetMock.uuid } returns "test"
-        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns mockk<ConnectionConfig>()
+        val connectionConfig = mockk<ConnectionConfig>()
+        every { connectionConfig.uuid } returns "fake_uuid"
+        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns connectionConfig
         val dataOpsManager = DataOpsManager.getService() as TestDataOpsManagerImpl
         dataOpsManager.testInstance = object : TestDataOpsManagerImpl() {
           override fun <R : Any> performOperation(operation: Operation<R>, progressIndicator: ProgressIndicator): R {
@@ -522,7 +532,9 @@ class AllocateDatasetActionTestSpec : WithApplicationShouldSpec({
         every { viewMock.mySelectedNodesData } returns selectedNodesData
         every { workingSetMock.name } returns "test"
         every { workingSetMock.uuid } returns "test"
-        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns mockk<ConnectionConfig>()
+        val connectionConfig = mockk<ConnectionConfig>()
+        every { connectionConfig.uuid } returns "fake_uuid"
+        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns connectionConfig
         val dataOpsManager = DataOpsManager.getService() as TestDataOpsManagerImpl
         dataOpsManager.testInstance = object : TestDataOpsManagerImpl() {
           override fun <R : Any> performOperation(operation: Operation<R>, progressIndicator: ProgressIndicator): R {
@@ -574,7 +586,9 @@ class AllocateDatasetActionTestSpec : WithApplicationShouldSpec({
 
         every { nodeMock.hint(FilesWorkingSet::class).unit } returns workingSetMock
         every { viewMock.mySelectedNodesData } returns selectedNodesData
-        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns mockk<ConnectionConfig>()
+        val connectionConfig = mockk<ConnectionConfig>()
+        every { connectionConfig.uuid } returns "fake_uuid"
+        every { workingSetMock.hint(ConnectionConfig::class).connectionConfig } returns connectionConfig
         val dataOpsManager = DataOpsManager.getService() as TestDataOpsManagerImpl
         dataOpsManager.testInstance = object : TestDataOpsManagerImpl() {
           override fun <R : Any> performOperation(operation: Operation<R>, progressIndicator: ProgressIndicator): R {